+++ packages/assemble-release-plan/src/utils.ts
@@ -1,4 +1,4 @@
-import { PackageGroup, VersionType } from "@changesets/types";
+import { PackageGroup, VersionType } from "@abizzle/changesets-types";
 import { Package } from "@manypkg/get-packages";
 import semver from "semver";
 import { InternalRelease } from "./types";