+++ packages/changelog-github/src/index.ts
@@ -1,4 +1,4 @@
-import { ChangelogFunctions } from "@changesets/types";
+import { ChangelogFunctions } from "@abizzle/changesets-types";
 // @ts-ignore
 import { config } from "dotenv";
 import { getInfo, getInfoFromPullRequest } from "@changesets/get-github-info";
@@ -13,7 +13,7 @@ const changelogFunctions: ChangelogFunctions = {
   ) => {
     if (!options.repo) {
       throw new Error(
-        'Please provide a repo to this changelog generator like this:\n"changelog": ["@changesets/changelog-github", { "repo": "org/repo" }]'
+        'Please provide a repo to this changelog generator like this:\n"changelog": ["@abizzle/changesets-changelog-github", { "repo": "org/repo" }]'
       );
     }
     if (dependenciesUpdated.length === 0) return "";
@@ -43,7 +43,7 @@ const changelogFunctions: ChangelogFunctions = {
   getReleaseLine: async (changeset, type, options) => {
     if (!options || !options.repo) {
       throw new Error(
-        'Please provide a repo to this changelog generator like this:\n"changelog": ["@changesets/changelog-github", { "repo": "org/repo" }]'
+        'Please provide a repo to this changelog generator like this:\n"changelog": ["@abizzle/changesets-changelog-github", { "repo": "org/repo" }]'
       );
     }
 
@@ -112,6 +112,9 @@ const changelogFunctions: ChangelogFunctions = {
     const prefix = [
       links.pull === null ? "" : ` ${links.pull}`,
       links.commit === null ? "" : ` ${links.commit}`,
+      changeset.groupedChangelog
+        ? ` **(${changeset.releases.map(({ name }) => name).join(", ")})**`
+        : "",
       users === null ? "" : ` Thanks ${users}!`,
     ].join("");