+++ packages/cli/src/commands/status/index.ts
@@ -4,14 +4,14 @@ import fs from "fs-extra";
 import path from "path";
 
 import * as git from "@changesets/git";
-import getReleasePlan from "@changesets/get-release-plan";
+import getReleasePlan from "@abizzle/changesets-get-release-plan";
 import { error, log, info, warn } from "@changesets/logger";
 import {
   VersionType,
   Release,
   ComprehensiveRelease,
   Config,
-} from "@changesets/types";
+} from "@abizzle/changesets-types";
 
 export default async function getStatus(
   cwd: string,
@@ -37,7 +37,18 @@ export default async function getStatus(
   const sinceBranch =
     since === undefined ? (sinceMaster ? "master" : undefined) : since;
   const releasePlan = await getReleasePlan(cwd, sinceBranch, config);
-  const { changesets, releases } = releasePlan;
+  const { changesets, individualReleases, groupedReleases } = releasePlan;
+  const releases = individualReleases.concat(
+    ...groupedReleases.flatMap<ComprehensiveRelease>((g) =>
+      g.projects.map((p) => ({
+        changesets: g.changesets,
+        name: p.name,
+        newVersion: g.newVersion,
+        oldVersion: p.oldVersion,
+        type: p.type,
+      }))
+    )
+  );
   const changedPackages = await git.getChangedPackagesSinceRef({
     cwd,
     ref: sinceBranch || config.baseBranch,