+++ packages/cli/src/commands/version/index.ts
@@ -2,10 +2,10 @@ import chalk from "chalk";
 import path from "path";
 import * as git from "@changesets/git";
 import { log, warn, error } from "@changesets/logger";
-import { Config } from "@changesets/types";
-import applyReleasePlan from "@changesets/apply-release-plan";
+import { Config } from "@abizzle/changesets-types";
+import applyReleasePlan from "@abizzle/changesets-apply-release-plan";
 import readChangesets from "@changesets/read";
-import assembleReleasePlan from "@changesets/assemble-release-plan";
+import assembleReleasePlan from "@abizzle/changesets-assemble-release-plan";
 import { getPackages } from "@manypkg/get-packages";
 
 import { removeEmptyFolders } from "../../utils/v1-legacy/removeFolders";